text_service/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.3
redis[hiredis]==5.2.0
pydantic==2.11.7
openai==1.55.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
